Move Thermal in file
